cleaned the app.py
